Move server startup to a package so integration tests can use it.
